Folder Restructuring
* Restructuring folder and file paths
* Categorizing for better lookups
